requirements.txt
## The following requirements were added by pip freeze:
torchinfo[pytorch]==1.6.2
